requirements.txt
streamlit==0.69.2
pillow==7.1.2 
tensorflow==2.2.0 
